Refactor scene
