+++ src/base/scene/mod.rs
@@ -1,0 +1,13 @@
+
+mod asset;
+mod material;
+mod aabb;
+mod procedural;
+mod primitive;
+mod mesh;
+mod buffer;
+mod scene;
+
+pub use scene::Scene;
+pub use asset::SceneAsset;
+pub use scene::SceneBuilder;
